requirements/dev.txt
-r prod.txt

# Testing
mock
pytest
pytest-flask
factory-boy

# Management script
Flask-Script